+++ Extra Profiles and Skill Sets/lua/multiprofileitemgui.lua
@@ -1,5 +1,5 @@
---30 character profile names
-Hooks:PostHook(MultiProfileItemGui, "init", "fragProfiles_post_MultiProfileItemGui_init", function(self, ...)
+--More characters for profile names
+Hooks:PostHook(MultiProfileItemGui, "init", "EPSS-PostHook-MultiProfileItemGui:init", function(self, ...)
 	self._max_length = 30
 	self:update()
 end)