+++ Extra Profiles and Skill Sets/lua/skilltreetweakdata.lua
@@ -1,11 +1,17 @@
-dofile(ModPath .. "lua/setup.lua")
-
 --Add more skill sets
---Add skill sets until table length equals total_profiles, loop does nothing if table length >= total_profiles
---So if the number of skill sets in the base game is changed, the added skill sets will not be lost
-Hooks:PostHook(SkillTreeTweakData, "init", "fragProfiles_post_SkillTreeTweakData_init", function(self, ...)
-	local actual_skill_sets = #self.skill_switches
-	for i=actual_skill_sets+1,fragProfiles._settings.total_profiles do
-		table.insert(self.skill_switches, {name_id = "fragProfiles_set_"..tostring(i)})
+--Make sure our setting is not lower than the number in the base game, update if necessary
+Hooks:PostHook(SkillTreeTweakData, "init", "EPSS-PostHook-SkillTreeTweakData:init", function(self, ...)
+	--Get the base game number of profiles
+	--EPSS updates number of profiles if necessary
+	local base_num_profiles = #self.skill_switches
+	EPSS:update_session_settings(base_num_profiles)
+
+	local start_idx = #self.skill_switches + 1
+	for i = start_idx, EPSS._session_profiles do
+		table.insert(self.skill_switches, {})
+	end
+
+	if EPSS.settings.allow_fewer then
+		table.crop(self.skill_switches, EPSS._session_profiles)
 	end
 end)