+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v2.0
+
+*2026-02-08 - Update 242.2*
+
+- Minimum number of profiles is automatically set to base game value (currently 30).
+- Default number of profiles increased to 45 (your current settings will not be affected).
+- Maximum number of profiles increased to 120.
+- All skill sets are unlocked automatically (including base game ones).
+- Added an option to automatically switch to the corresponding skill set (if not suspended) when equipping a profile. Enabled by default.
+- Added an option to lower the minimum number of profiles below the base game value. Disabled by default.
+
 ## v1.0
 
 *2020-02-28 - Update 199 Mark II*